+++ packages/terra-functional-testing/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added configuration to retrieve and unzip screenshots from registry.
+
 ## 1.5.0 - (May 19, 2021)
 
 * Removed